Move arXiv tests to a separate file
